refactor(onboarding): componentize + M3 unify, split god file
Split OnboardingScreen.kt (1988 lines) into per-step files under
ui/screens/onboarding/ (Welcome, Density, ContentFilters, Permissions,
BottomNavSetup); the orchestrator keeps only the step enum + flow.
LoadingScreen + DisposableLifecycle move to a shared ui/screens file.

Unify every onboarding button on three shared M3 components
(OnboardingPrimaryButton / OnboardingActionButton / OnboardingTextButton) so
shape, ripple and disabled state stop drifting (was a mix of 8/9/10/12dp
hand-rolled shapes). Convert the three hand-rolled Box+Surface scrim dialogs
(restart, custom density, legal) to the shared DefaultDialog.

Replace every alpha-blended card fill (surfaceVariant/primaryContainer .copy)
with one shared onboardingCardColors() helper: solid secondaryContainer when
active, surfaceContainerHigh otherwise (same neutral tone as the search field).
Neutral step titles via OnboardingStepTitle; drop the remaining pink accent
text and the no-op Welcome gradient. Tidy imports and split-artifact indents.
