+++ app/src/main/kotlin/com/jtech/zemer/ui/screens/OnboardingScreen.kt
@@ -1,207 +1,27 @@
 package com.jtech.zemer.ui.screens
-import com.jtech.zemer.ui.component.AppNameTitle
 
-import android.Manifest
-import android.annotation.SuppressLint
 import android.content.Context
-import android.content.Intent
-import android.graphics.PorterDuff
-import android.graphics.PorterDuffColorFilter
-import android.net.Uri
-import android.os.Build
-import android.os.PowerManager
-import android.provider.Settings
-import androidx.core.content.edit
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.contract.ActivityResultContracts
-import com.google.android.gms.auth.api.signin.GoogleSignIn
-import com.google.android.gms.common.api.ApiException
-import androidx.compose.animation.animateColorAsState
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
-import androidx.compose.material3.RadioButton
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.foundation.layout.width
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.core.app.NotificationManagerCompat
-import androidx.core.content.ContextCompat
-import androidx.core.net.toUri
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleEventObserver
-import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import com.jtech.zemer.viewmodels.OnboardingViewModel
-import com.airbnb.lottie.LottieProperty
-import com.airbnb.lottie.compose.LottieAnimation
-import com.airbnb.lottie.compose.LottieClipSpec
-import com.airbnb.lottie.compose.LottieCompositionSpec
-import com.airbnb.lottie.compose.LottieConstants
-import com.airbnb.lottie.compose.animateLottieCompositionAsState
-import com.airbnb.lottie.compose.rememberLottieComposition
-import com.airbnb.lottie.compose.rememberLottieDynamicProperties
-import com.airbnb.lottie.compose.rememberLottieDynamicProperty
-import com.jtech.zemer.R
-import com.jtech.zemer.ui.component.SyncAccountWarning
-import com.jtech.zemer.ui.component.DefaultDialog
-import com.jtech.zemer.ui.screens.onboarding.OnboardingChoiceCard
 import com.jtech.zemer.ui.screens.onboarding.OnboardingSearchBackupScreen
-import com.jtech.zemer.constants.DensityScale
-import com.jtech.zemer.utils.PermissionHelper
-import com.jtech.zemer.extensions.isInternetConnected
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.core.booleanPreferencesKey
-import androidx.datastore.preferences.core.edit
 import com.jtech.zemer.constants.EnableContentFiltersKey
 import com.jtech.zemer.constants.AllowFemaleSingersKey
 import com.jtech.zemer.constants.BlockVideosKey
-import com.jtech.zemer.utils.ContentFilterConfig
-import com.jtech.zemer.utils.rememberPreference
-import com.jtech.zemer.di.MainDataStore
-import dagger.hilt.android.qualifiers.ApplicationContext
-import kotlinx.coroutines.flow.first
-import com.google.firebase.auth.FirebaseAuth
-import kotlinx.coroutines.launch
-import androidx.compose.material3.Switch
-import androidx.compose.material3.SwitchDefaults
-import androidx.compose.ui.res.painterResource
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.delay
+import com.jtech.zemer.ui.screens.onboarding.WelcomeScreen
+import com.jtech.zemer.ui.screens.onboarding.DensityScreen
+import com.jtech.zemer.ui.screens.onboarding.ContentFiltersScreen
+import com.jtech.zemer.ui.screens.onboarding.PermissionsScreen
+import com.jtech.zemer.ui.screens.onboarding.BottomNavSetupScreen
 
 private enum class OnboardingStep { Welcome, Density, ContentFilters, Permissions, BottomNavSetup, SearchBackup, Loading }
-private enum class LegalKind { TOS, PRIVACY }
-
-@Composable
-private fun NetworkStatusBanner(
-    modifier: Modifier = Modifier
-) {
-    val context = LocalContext.current
-    var isConnected by remember { mutableStateOf(false) }
-    var isChecking by remember { mutableStateOf(true) }
-
-    // Periodic network check on background thread
-    LaunchedEffect(Unit) {
-        // Initial check
-        isConnected = withContext(Dispatchers.IO) {
-            context.isInternetConnected()
-        }
-        isChecking = false
-
-        // Continue checking every 2 seconds
-        while (true) {
-            delay(2000)
-            val newConnectionState = withContext(Dispatchers.IO) {
-                context.isInternetConnected()
-            }
-            if (newConnectionState != isConnected) {
-                isConnected = newConnectionState
-            }
-        }
-    }
-
-    Card(
-        modifier = modifier.fillMaxWidth(),
-        colors = CardDefaults.cardColors(
-            containerColor = when {
-                isChecking -> MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.5f)
-                isConnected -> MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.3f)
-                else -> MaterialTheme.colorScheme.errorContainer.copy(alpha = 0.3f)
-            }
-        ),
-        shape = RoundedCornerShape(8.dp)
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(12.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            if (isChecking) {
-                CircularProgressIndicator(
-                    modifier = Modifier.size(16.dp),
-                    strokeWidth = 2.dp,
-                    color = MaterialTheme.colorScheme.primary
-                )
-            } else {
-                Icon(
-                    painter = painterResource(
-                        id = if (isConnected) R.drawable.wifi_proxy else R.drawable.offline
-                    ),
-                    contentDescription = null,
-                    modifier = Modifier.size(20.dp),
-                    tint = if (isConnected)
-                        MaterialTheme.colorScheme.primary
-                    else
-                        MaterialTheme.colorScheme.error
-                )
-            }
-
-            Text(
-                text = when {
-                    isChecking -> stringResource(R.string.network_checking_connection)
-                    isConnected -> stringResource(R.string.network_connected)
-                    else -> stringResource(R.string.network_no_connection)
-                },
-                style = MaterialTheme.typography.bodySmall,
-                color = when {
-                    isChecking -> MaterialTheme.colorScheme.onSurfaceVariant
-                    isConnected -> MaterialTheme.colorScheme.primary
-                    else -> MaterialTheme.colorScheme.error
-                }
-            )
-        }
-    }
-}
 
 @Composable
 fun OnboardingFlow(
@@ -273,1716 +93,3 @@ fun OnboardingFlow(
     }
 }
 
-@Composable
-private fun WelcomeScreen(
-    onContinue: () -> Unit,
-) {
-    val context = LocalContext.current
-    val scope = rememberCoroutineScope()
-    val gradient = Brush.verticalGradient(
-        listOf(
-            MaterialTheme.colorScheme.surface,
-            MaterialTheme.colorScheme.surface
-        )
-    )
-    var legal by remember { mutableStateOf<LegalKind?>(null) }
-    var agreed by rememberSaveable { mutableStateOf(false) }
-    var isConnected by remember { mutableStateOf(false) }
-    var isCheckingNetwork by remember { mutableStateOf(true) }
-
-    // Initial and periodic network checks
-    LaunchedEffect(Unit) {
-        // Initial check
-        isConnected = withContext(Dispatchers.IO) {
-            context.isInternetConnected()
-        }
-        isCheckingNetwork = false
-
-        // Continue checking every 2 seconds
-        while (true) {
-            delay(2000)
-            val newConnectionState = withContext(Dispatchers.IO) {
-                context.isInternetConnected()
-            }
-            if (newConnectionState != isConnected) {
-                isConnected = newConnectionState
-            }
-        }
-    }
-
-    val composition by rememberLottieComposition(LottieCompositionSpec.RawRes(R.raw.welcome))
-    val animationState = animateLottieCompositionAsState(
-        composition = composition,
-        iterations = 1,
-        clipSpec = LottieClipSpec.Progress(0f, 0.5f),
-        restartOnPlay = false
-    )
-    val lottieColors = rememberLottieDynamicProperties(
-        rememberLottieDynamicProperty(
-            property = LottieProperty.COLOR_FILTER,
-            value = PorterDuffColorFilter(MaterialTheme.colorScheme.primary.toArgb(), PorterDuff.Mode.SRC_ATOP),
-            keyPath = arrayOf("**")
-        )
-    )
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(gradient),
-        contentAlignment = Alignment.Center
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth(0.9f)
-                .fillMaxHeight()
-                .padding(vertical = 48.dp),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Column(
-                verticalArrangement = Arrangement.spacedBy(12.dp),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                LottieAnimation(
-                    composition = composition,
-                    progress = { animationState.progress },
-                    dynamicProperties = lottieColors,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 24.dp, vertical = 8.dp)
-                )
-                AppNameTitle(textAlign = TextAlign.Center)
-            }
-
-            Spacer(modifier = Modifier.weight(1f))
-
-            Column(
-                verticalArrangement = Arrangement.spacedBy(12.dp),
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(bottom = 12.dp)
-            ) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(12.dp),
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clickable { agreed = !agreed }
-                ) {
-                    Checkbox(checked = agreed, onCheckedChange = { agreed = it })
-                    Column(
-                        modifier = Modifier.weight(1f),
-                        verticalArrangement = Arrangement.spacedBy(6.dp)
-                    ) {
-                        Text(
-                            text = stringResource(R.string.onboarding_agree_label),
-                            style = MaterialTheme.typography.bodySmall.copy(fontWeight = FontWeight.Medium),
-                            color = MaterialTheme.colorScheme.onSurface
-                        )
-                        Row(
-                            horizontalArrangement = Arrangement.spacedBy(12.dp),
-                            modifier = Modifier.padding(top = 2.dp)
-                        ) {
-                            Text(
-                                text = stringResource(R.string.onboarding_view_tos),
-                                color = MaterialTheme.colorScheme.primary,
-                                style = MaterialTheme.typography.labelSmall,
-                                modifier = Modifier.clickable { legal = LegalKind.TOS }
-                            )
-                            Text(
-                                text = "•",
-                                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                                style = MaterialTheme.typography.labelSmall
-                            )
-                            Text(
-                                text = stringResource(R.string.onboarding_view_privacy),
-                                color = MaterialTheme.colorScheme.primary,
-                                style = MaterialTheme.typography.labelSmall,
-                                modifier = Modifier.clickable { legal = LegalKind.PRIVACY }
-                            )
-                        }
-                    }
-                }
-
-                Button(
-                    onClick = onContinue,
-                    enabled = agreed && isConnected && !isCheckingNetwork,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(48.dp),
-                    shape = RoundedCornerShape(12.dp)
-                ) {
-                    if (isCheckingNetwork) {
-                        CircularProgressIndicator(
-                            modifier = Modifier.size(16.dp),
-                            strokeWidth = 2.dp,
-                            color = MaterialTheme.colorScheme.onPrimary
-                        )
-                    } else {
-                        Text(
-                            text = when {
-                                !isConnected -> stringResource(R.string.network_internet_required)
-                                else -> stringResource(R.string.onboarding_continue)
-                            },
-                            style = MaterialTheme.typography.labelMedium
-                        )
-                    }
-                }
-
-                if (!isConnected || isCheckingNetwork) {
-                    Text(
-                        text = when {
-                            isCheckingNetwork -> stringResource(R.string.network_checking_internet)
-                            !isConnected -> stringResource(R.string.network_required_setup)
-                            else -> ""
-                        },
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        textAlign = TextAlign.Center,
-                        modifier = Modifier.fillMaxWidth()
-                    )
-                }
-            }
-        }
-
-        legal?.let { kind ->
-            LegalOverlay(
-                title = stringResource(
-                    if (kind == LegalKind.TOS) R.string.onboarding_tos_title else R.string.onboarding_privacy_title
-                ),
-                body = stringResource(
-                    // The telemetry variant additionally discloses the anonymous usage tracking —
-                    // the older strings falsely claim "no usage analytics" once feat/track ships.
-                    if (kind == LegalKind.TOS) R.string.onboarding_tos_body else R.string.onboarding_privacy_body_telemetry
-                ),
-                onDismiss = { legal = null }
-            )
-        }
-    }
-}
-
-@Composable
-private fun DensityScreen(
-    onSkip: () -> Unit,
-    onBack: () -> Unit,
-) {
-    val context = LocalContext.current
-    var isConnected by remember { mutableStateOf(false) }
-    var isCheckingNetwork by remember { mutableStateOf(true) }
-
-    // Initial and periodic network checks
-    LaunchedEffect(Unit) {
-        // Initial check
-        isConnected = withContext(Dispatchers.IO) {
-            context.isInternetConnected()
-        }
-        isCheckingNetwork = false
-
-        // Continue checking every 2 seconds
-        while (true) {
-            delay(2000)
-            val newConnectionState = withContext(Dispatchers.IO) {
-                context.isInternetConnected()
-            }
-            if (newConnectionState != isConnected) {
-                isConnected = newConnectionState
-            }
-        }
-    }
-    var selectedDensity by rememberSaveable { mutableStateOf(DensityScale.NATIVE) }
-    var customDensityValue by rememberSaveable { mutableStateOf(0.85f) }
-    var showRestartDialog by rememberSaveable { mutableStateOf(false) }
-    var showCustomDensityDialog by rememberSaveable { mutableStateOf(false) }
-
-    val densityOptions = DensityScale.entries
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(MaterialTheme.colorScheme.surface)
-    ) {
-        Column(
-            verticalArrangement = Arrangement.spacedBy(8.dp),
-            modifier = Modifier
-                .align(Alignment.Center)
-                .fillMaxWidth(0.92f)
-        ) {
-            
-            Column(
-                verticalArrangement = Arrangement.spacedBy(2.dp),
-                modifier = Modifier.fillMaxWidth(),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Text(
-                    text = stringResource(R.string.onboarding_density_title),
-                    style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.Bold),
-                    textAlign = TextAlign.Center,
-                    color = MaterialTheme.colorScheme.primary
-                )
-                Text(
-                    text = stringResource(R.string.onboarding_density_subtitle),
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    textAlign = TextAlign.Center
-                )
-            }
-
-            Spacer(modifier = Modifier.height(8.dp))
-
-            Card(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .border(
-                        width = 1.dp,
-                        color = MaterialTheme.colorScheme.primary.copy(alpha = 0.3f),
-                        shape = RoundedCornerShape(10.dp)
-                    ),
-                colors = CardDefaults.cardColors(
-                    containerColor = MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.5f)
-                ),
-                shape = RoundedCornerShape(10.dp)
-            ) {
-                Column(modifier = Modifier.padding(4.dp)) {
-                    densityOptions.forEach { density ->
-                        Row(
-                            verticalAlignment = Alignment.CenterVertically,
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .clip(RoundedCornerShape(6.dp))
-                                .clickable {
-                                    if (density == DensityScale.CUSTOM) {
-                                        showCustomDensityDialog = true
-                                    } else {
-                                        selectedDensity = density
-                                    }
-                                }
-                                .padding(horizontal = 4.dp, vertical = 6.dp)
-                        ) {
-                            RadioButton(
-                                selected = selectedDensity == density,
-                                onClick = {
-                                    if (density == DensityScale.CUSTOM) {
-                                        showCustomDensityDialog = true
-                                    } else {
-                                        selectedDensity = density
-                                    }
-                                },
-                                modifier = Modifier.size(36.dp)
-                            )
-                            Text(
-                                text = if (density == DensityScale.CUSTOM && selectedDensity == DensityScale.CUSTOM) {
-                                    stringResource(R.string.density_label_custom_value, (customDensityValue * 100).toInt())
-                                } else {
-                                    stringResource(density.labelRes)
-                                },
-                                style = MaterialTheme.typography.bodySmall.copy(fontWeight = FontWeight.Medium),
-                                color = MaterialTheme.colorScheme.onSurface,
-                                modifier = Modifier.padding(start = 4.dp)
-                            )
-                        }
-                    }
-                }
-            }
-
-            Spacer(Modifier.height(8.dp))
-
-            Column(
-                modifier = Modifier.fillMaxWidth(),
-                verticalArrangement = Arrangement.spacedBy(6.dp)
-            ) {
-                if (selectedDensity != DensityScale.NATIVE) {
-                    Button(
-                        onClick = {
-                            val densityValue = if (selectedDensity == DensityScale.CUSTOM) {
-                                customDensityValue
-                            } else {
-                                selectedDensity.value
-                            }
-                            context.getSharedPreferences("metrolist_settings", Context.MODE_PRIVATE)
-                                .edit {
-                                    putFloat("density_scale_factor", densityValue)
-                                }
-                            showRestartDialog = true
-                        },
-                        enabled = isConnected && !isCheckingNetwork,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .height(40.dp),
-                        shape = RoundedCornerShape(8.dp),
-                        colors = ButtonDefaults.buttonColors(
-                            containerColor = MaterialTheme.colorScheme.primary
-                        )
-                    ) {
-                        Text(
-                            text = if (isCheckingNetwork) stringResource(R.string.network_checking)
-                                  else if (!isConnected) stringResource(R.string.network_internet_required)
-                                  else stringResource(R.string.onboarding_apply_density),
-                            style = MaterialTheme.typography.labelMedium
-                        )
-                    }
-                }
-
-                OutlinedButton(
-                    onClick = onSkip,
-                    enabled = isConnected && !isCheckingNetwork,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(40.dp),
-                    shape = RoundedCornerShape(8.dp),
-                    border = BorderStroke(1.dp, if (isConnected && !isCheckingNetwork) MaterialTheme.colorScheme.outline else MaterialTheme.colorScheme.outline.copy(alpha = 0.3f))
-                ) {
-                    Text(
-                        text = if (isCheckingNetwork) stringResource(R.string.network_checking)
-                              else if (!isConnected) stringResource(R.string.network_internet_required)
-                              else stringResource(R.string.onboarding_skip),
-                        style = MaterialTheme.typography.labelMedium,
-                        color = if (isConnected && !isCheckingNetwork) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-
-                if (!isConnected || isCheckingNetwork) {
-                    Text(
-                        text = if (isCheckingNetwork) stringResource(R.string.network_checking_internet)
-                              else stringResource(R.string.network_required_to_continue),
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        textAlign = TextAlign.Center,
-                        modifier = Modifier.fillMaxWidth()
-                    )
-                }
-
-                TextButton(
-                    onClick = onBack,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(32.dp)
-                ) {
-                    Text(
-                        text = stringResource(R.string.onboarding_back),
-                        style = MaterialTheme.typography.labelSmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-            }
-        }
-    }
-
-    if (showCustomDensityDialog) {
-        CustomDensityDialog(
-            initialValue = customDensityValue,
-            onDismiss = { showCustomDensityDialog = false },
-            onConfirm = { value ->
-                customDensityValue = value
-                selectedDensity = DensityScale.CUSTOM
-                showCustomDensityDialog = false
-            }
-        )
-    }
-
-    if (showRestartDialog) {
-        RestartDialog(
-            onDismiss = { showRestartDialog = false },
-            onRestart = {
-                showRestartDialog = false
-                restartApp(context)
-            }
-        )
-    }
-}
-
-@Composable
-private fun RestartDialog(
-    onDismiss: () -> Unit,
-    onRestart: () -> Unit,
-) {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.Black.copy(alpha = 0.4f))
-            .clickable { onDismiss() },
-        contentAlignment = Alignment.Center
-    ) {
-        Surface(
-            modifier = Modifier
-                .fillMaxWidth(0.88f)
-                .padding(20.dp),
-            shape = RoundedCornerShape(16.dp),
-            tonalElevation = 6.dp,
-            color = MaterialTheme.colorScheme.surface
-        ) {
-            Column(
-                modifier = Modifier.padding(20.dp),
-                verticalArrangement = Arrangement.spacedBy(12.dp)
-            ) {
-                Text(
-                    text = stringResource(R.string.restart_required),
-                    style = MaterialTheme.typography.titleLarge.copy(fontWeight = FontWeight.Bold),
-                    color = MaterialTheme.colorScheme.primary
-                )
-                Text(
-                    text = stringResource(R.string.density_restart_message),
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.End,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    TextButton(onClick = onDismiss) {
-                        Text(stringResource(android.R.string.cancel))
-                    }
-                    Button(
-                        onClick = onRestart,
-                        modifier = Modifier.padding(start = 8.dp),
-                        shape = RoundedCornerShape(10.dp)
-                    ) {
-                        Text(stringResource(R.string.restart))
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun CustomDensityDialog(
-    initialValue: Float,
-    onDismiss: () -> Unit,
-    onConfirm: (Float) -> Unit,
-) {
-    var textValue by remember { mutableStateOf((initialValue * 100).toInt().toString()) }
-    var isError by remember { mutableStateOf(false) }
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.Black.copy(alpha = 0.4f))
-            .clickable { onDismiss() },
-        contentAlignment = Alignment.Center
-    ) {
-        Surface(
-            modifier = Modifier
-                .fillMaxWidth(0.88f)
-                .padding(20.dp)
-                .clickable(enabled = false) { },
-            shape = RoundedCornerShape(16.dp),
-            tonalElevation = 6.dp,
-            color = MaterialTheme.colorScheme.surface
-        ) {
-            Column(
-                modifier = Modifier.padding(20.dp),
-                verticalArrangement = Arrangement.spacedBy(12.dp)
-            ) {
-                Text(
-                    text = stringResource(R.string.custom_density_title),
-                    style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.Bold),
-                    color = MaterialTheme.colorScheme.primary
-                )
-                Text(
-                    text = stringResource(R.string.custom_density_hint),
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-                androidx.compose.material3.OutlinedTextField(
-                    value = textValue,
-                    onValueChange = { newValue ->
-                        textValue = newValue.filter { it.isDigit() }
-                        val intValue = textValue.toIntOrNull()
-                        isError = intValue == null || intValue !in 50..120
-                    },
-                    label = { Text("%") },
-                    isError = isError,
-                    singleLine = true,
-                    modifier = Modifier.fillMaxWidth()
-                )
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.End,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    TextButton(onClick = onDismiss) {
-                        Text(stringResource(android.R.string.cancel))
-                    }
-                    Button(
-                        onClick = {
-                            val intValue = textValue.toIntOrNull()
-                            if (intValue != null && intValue in 50..120) {
-                                onConfirm(intValue / 100f)
-                            }
-                        },
-                        enabled = !isError && textValue.isNotEmpty(),
-                        modifier = Modifier.padding(start = 8.dp),
-                        shape = RoundedCornerShape(10.dp)
-                    ) {
-                        Text(stringResource(R.string.ok))
-                    }
-                }
-            }
-        }
-    }
-}
-
-private fun restartApp(context: Context) {
-    val intent = context.packageManager.getLaunchIntentForPackage(context.packageName)?.apply {
-        addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK)
-    }
-    context.startActivity(intent)
-    Runtime.getRuntime().exit(0)
-}
-
-@Composable
-private fun ContentFiltersScreen(
-    onBack: () -> Unit,
-    onSkip: () -> Unit,
-    viewModel: OnboardingViewModel = hiltViewModel(),
-    contentFiltersAlreadySet: Boolean = false
-) {
-    val context = LocalContext.current
-    val scope = rememberCoroutineScope()
-    val uiState by viewModel.uiState.collectAsState()
-    val authState by viewModel.authState.collectAsState(initial = com.jtech.zemer.auth.AuthState.SignedOut)
-    var isConnected by remember { mutableStateOf(false) }
-    var isCheckingNetwork by remember { mutableStateOf(true) }
-
-    // Initial and periodic network checks
-    LaunchedEffect(Unit) {
-        // Initial check
-        isConnected = withContext(Dispatchers.IO) {
-            context.isInternetConnected()
-        }
-        isCheckingNetwork = false
-
-        // Continue checking every 2 seconds
-        while (true) {
-            delay(2000)
-            val newConnectionState = withContext(Dispatchers.IO) {
-                context.isInternetConnected()
-            }
-            if (newConnectionState != isConnected) {
-                isConnected = newConnectionState
-            }
-        }
-    }
-
-    // Content filter states (using rememberPreference to auto-save to DataStore)
-    val (enableContentFilters, onEnableContentFiltersChange) = rememberPreference(key = EnableContentFiltersKey, defaultValue = true)
-    val (allowFemaleSingers, onAllowFemaleSingersChange) = rememberPreference(key = AllowFemaleSingersKey, defaultValue = false)
-    val (blockVideos, onBlockVideosChange) = rememberPreference(key = BlockVideosKey, defaultValue = false)
-    var signInDelaySeconds by remember { mutableStateOf(0) }
-    var showSignInDialog by remember { mutableStateOf(false) }
-
-    // Google Sign-In launcher (matching ContentSettings implementation)
-    val googleSignInLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.StartActivityForResult()
-    ) { result ->
-        if (result.resultCode == android.app.Activity.RESULT_OK) {
-            // Handle Google Sign-In result
-            val data = result.data
-            val task = GoogleSignIn.getSignedInAccountFromIntent(data)
-            try {
-                val account = task.getResult(ApiException::class.java)
-                val idToken = account.idToken
-                if (idToken != null) {
-                    scope.launch {
-                        viewModel.signInWithGoogle(idToken)
-                        showSignInDialog = false
-                    }
-                }
-            } catch (e: ApiException) {
-                showSignInDialog = false
-                // Handle sign-in failure
-            }
-        } else {
-            showSignInDialog = false
-        }
-    }
-
-    // Check for auto-restore and skip logic
-    LaunchedEffect(Unit) {
-        // Check if we should skip (already configured)
-        if (contentFiltersAlreadySet) {
-            onSkip()
-            return@LaunchedEffect
-        }
-
-        // Attempt auto-restore only if user not signed in
-        if (authState !is com.jtech.zemer.auth.AuthState.SignedIn) {
-            viewModel.attemptAutoRestore()
-        }
-    }
-
-    // Update UI state when server preferences are found
-    LaunchedEffect(uiState.restoredConfig) {
-        uiState.restoredConfig?.let { config ->
-            // Note: rememberPreference will automatically load the restored values
-            // No need to manually set them here
-
-            // Auto-proceed after showing restore UI for 3 seconds
-            kotlinx.coroutines.delay(3000)
-            onSkip()
-        }
-    }
-
-    // Handle countdown for sign-in delay
-    LaunchedEffect(showSignInDialog) {
-        if (showSignInDialog) {
-            signInDelaySeconds = 5
-            for (i in 5 downTo 1) {
-                kotlinx.coroutines.delay(1000)
-                signInDelaySeconds = i - 1
-            }
-        } else {
-            signInDelaySeconds = 0
-        }
-    }
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(MaterialTheme.colorScheme.surface)
-    ) {
-        Column(
-            verticalArrangement = Arrangement.spacedBy(18.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            modifier = Modifier
-                .align(Alignment.Center)
-                .fillMaxWidth(0.9f)
-                .verticalScroll(rememberScrollState())
-                .padding(horizontal = 16.dp, vertical = 24.dp)
-        ) {
-                        if (uiState.isCheckingAutoRestore) {
-                // Loading state while checking for auto-restore
-                Column(
-                    verticalArrangement = Arrangement.spacedBy(16.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    Text(
-                        text = stringResource(R.string.onboarding_restoring_filters),
-                        style = MaterialTheme.typography.headlineSmall.copy(fontWeight = FontWeight.Bold),
-                        textAlign = TextAlign.Center,
-                        color = MaterialTheme.colorScheme.primary
-                    )
-                    CircularProgressIndicator(
-                        modifier = Modifier.size(48.dp),
-                        color = MaterialTheme.colorScheme.primary
-                    )
-                    Text(
-                        text = stringResource(R.string.onboarding_checking_saved),
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        textAlign = TextAlign.Center
-                    )
-                }
-            } else if (uiState.hasServerPreferences && uiState.restoredConfig != null) {
-                // Auto-restore screen showing what's being restored
-                Column(
-                    verticalArrangement = Arrangement.spacedBy(16.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    Text(
-                        text = stringResource(R.string.onboarding_restoring_filters),
-                        style = MaterialTheme.typography.headlineSmall.copy(fontWeight = FontWeight.Bold),
-                        textAlign = TextAlign.Center,
-                        color = MaterialTheme.colorScheme.primary
-                    )
-
-                    Text(
-                        text = stringResource(R.string.onboarding_found_saved),
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        textAlign = TextAlign.Center
-                    )
-
-                    // Show what's being restored
-                    Card(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .border(
-                                width = 1.5.dp,
-                                color = MaterialTheme.colorScheme.primary.copy(alpha = 0.3f),
-                                shape = RoundedCornerShape(12.dp)
-                            ),
-                        colors = CardDefaults.cardColors(
-                            containerColor = MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.3f)
-                        ),
-                        shape = RoundedCornerShape(12.dp)
-                    ) {
-                        Column(modifier = Modifier.padding(16.dp)) {
-                            val config = uiState.restoredConfig
-                            Text(
-                                text = stringResource(R.string.onboarding_restored_settings),
-                                style = MaterialTheme.typography.labelLarge.copy(fontWeight = FontWeight.SemiBold),
-                                color = MaterialTheme.colorScheme.onSurface,
-                                modifier = Modifier.padding(bottom = 8.dp)
-                            )
-
-                            Text(
-                                text = stringResource(R.string.onboarding_restored_filters_enabled),
-                                style = MaterialTheme.typography.bodyMedium,
-                                color = MaterialTheme.colorScheme.onSurface
-                            )
-
-                            Text(
-                                text = stringResource(R.string.onboarding_restored_female, stringResource(if (config?.allowFemaleSingers == true) R.string.allowed else R.string.blocked)),
-                                style = MaterialTheme.typography.bodyMedium,
-                                color = MaterialTheme.colorScheme.onSurface
-                            )
-
-                            Text(
-                                text = stringResource(R.string.onboarding_restored_videos, stringResource(if (config?.blockVideos == true) R.string.blocked else R.string.allowed)),
-                                style = MaterialTheme.typography.bodyMedium,
-                                color = MaterialTheme.colorScheme.onSurface
-                            )
-                        }
-                    }
-
-                    Text(
-                        text = stringResource(R.string.onboarding_continuing_auto),
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        textAlign = TextAlign.Center
-                    )
-                }
-            } else {
-                // Normal content filter setup screen
-                // Title and subtitle
-                Column(
-                    verticalArrangement = Arrangement.spacedBy(4.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    Text(
-                        text = stringResource(R.string.content_filters),
-                        style = MaterialTheme.typography.headlineSmall.copy(fontWeight = FontWeight.Bold),
-                        textAlign = TextAlign.Center,
-                        color = MaterialTheme.colorScheme.primary
-                    )
-                    Text(
-                        text = stringResource(R.string.onboarding_filters_subtitle),
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        textAlign = TextAlign.Center
-                    )
-                }
-
-                // Allow Female Singers toggle
-                FilterOptionCard(
-                    title = stringResource(R.string.onboarding_allow_female_title),
-                    description = stringResource(R.string.onboarding_allow_female_desc),
-                    isEnabled = allowFemaleSingers,
-                    onToggle = { onAllowFemaleSingersChange(it) },
-                    icon = R.drawable.person
-                )
-
-                // Block Videos toggle
-                FilterOptionCard(
-                    title = stringResource(R.string.onboarding_block_videos_title),
-                    description = stringResource(R.string.onboarding_block_videos_desc),
-                    isEnabled = blockVideos,
-                    onToggle = { onBlockVideosChange(it) },
-                    icon = R.drawable.ic_video_hd
-                )
-
-                // Sign-in status card
-                Card(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .border(
-                            width = 1.5.dp,
-                            color = MaterialTheme.colorScheme.primary.copy(alpha = 0.3f),
-                            shape = RoundedCornerShape(12.dp)
-                        ),
-                    colors = CardDefaults.cardColors(
-                        containerColor = if (authState is com.jtech.zemer.auth.AuthState.SignedIn)
-                            MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.4f)
-                        else
-                            MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.5f)
-                    ),
-                    shape = RoundedCornerShape(12.dp)
-                ) {
-                    Column(modifier = Modifier.padding(14.dp)) {
-                        Row(
-                            verticalAlignment = Alignment.CenterVertically,
-                            horizontalArrangement = Arrangement.SpaceBetween,
-                            modifier = Modifier.fillMaxWidth()
-                        ) {
-                            Column(modifier = Modifier.weight(1f)) {
-                                Text(
-                                    text = if (authState is com.jtech.zemer.auth.AuthState.SignedIn) stringResource(R.string.sync_account_created) else stringResource(R.string.sync_account_optional_create),
-                                    style = MaterialTheme.typography.labelLarge.copy(fontWeight = FontWeight.SemiBold),
-                                    color = MaterialTheme.colorScheme.onSurface
-                                )
-                                Text(
-                                    text = if (authState is com.jtech.zemer.auth.AuthState.SignedIn)
-                                        stringResource(R.string.sync_account_locked_backed_up)
-                                    else
-                                        stringResource(R.string.sync_account_connect_to_lock),
-                                    style = MaterialTheme.typography.labelSmall,
-                                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                                    modifier = Modifier.padding(top = 2.dp)
-                                )
-                            }
-                            Box(
-                                modifier = Modifier
-                                    .clip(CircleShape)
-                                    .background(
-                                        if (authState is com.jtech.zemer.auth.AuthState.SignedIn)
-                                            MaterialTheme.colorScheme.primary.copy(alpha = 0.12f)
-                                        else
-                                            MaterialTheme.colorScheme.outline.copy(alpha = 0.12f)
-                                    )
-                                    .padding(horizontal = 8.dp, vertical = 4.dp),
-                                contentAlignment = Alignment.Center
-                            ) {
-                                Text(
-                                    text = if (authState is com.jtech.zemer.auth.AuthState.SignedIn) stringResource(R.string.sync_account_active) else stringResource(R.string.sync_account_optional),
-                                    color = if (authState is com.jtech.zemer.auth.AuthState.SignedIn) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.outline,
-                                    style = MaterialTheme.typography.labelSmall.copy(fontWeight = FontWeight.Medium),
-                                    textAlign = TextAlign.Center
-                                )
-                            }
-                        }
-
-                        if (authState !is com.jtech.zemer.auth.AuthState.SignedIn) {
-                            Spacer(Modifier.height(10.dp))
-                            OutlinedButton(
-                                onClick = { showSignInDialog = true },
-                                modifier = Modifier.fillMaxWidth().height(38.dp),
-                                border = BorderStroke(1.dp, MaterialTheme.colorScheme.primary),
-                                colors = ButtonDefaults.outlinedButtonColors(
-                                    contentColor = MaterialTheme.colorScheme.primary
-                                )
-                            ) {
-                                Text(
-                                    stringResource(R.string.sync_account_create_title),
-                                    style = MaterialTheme.typography.labelSmall.copy(fontWeight = FontWeight.Medium)
-                                )
-                            }
-                        }
-                    }
-                }
-
-                // Action buttons
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.spacedBy(12.dp)
-                ) {
-                    OutlinedButton(
-                        onClick = onBack,
-                        modifier = Modifier.weight(1f).height(40.dp),
-                        border = BorderStroke(1.dp, MaterialTheme.colorScheme.outline),
-                        colors = ButtonDefaults.outlinedButtonColors(
-                            contentColor = MaterialTheme.colorScheme.onSurface
-                        )
-                    ) {
-                        Text(stringResource(R.string.onboarding_back))
-                    }
-
-                    Button(
-                        onClick = { onSkip() },
-                        enabled = isConnected && !isCheckingNetwork,
-                        modifier = Modifier.weight(2f).height(40.dp),
-                        colors = ButtonDefaults.buttonColors(
-                            containerColor = MaterialTheme.colorScheme.primary
-                        )
-                    ) {
-                        Text(if (isCheckingNetwork) stringResource(R.string.network_checking)
-                             else if (!isConnected) stringResource(R.string.network_internet_required)
-                             else stringResource(R.string.onboarding_continue))
-                    }
-                }
-            }
-        }
-
-        // Sign-in dialog (matching ContentSettings dialog)
-        if (showSignInDialog) {
-            DefaultDialog(
-                onDismiss = {
-                    showSignInDialog = false
-                    signInDelaySeconds = 0
-                },
-                horizontalAlignment = Alignment.Start,
-                title = { Text(stringResource(R.string.sync_account_important_title)) },
-                content = {
-                    SyncAccountWarning(
-                        delaySeconds = signInDelaySeconds,
-                        showCountdown = signInDelaySeconds > 0,
-                    )
-                },
-                buttons = {
-                    TextButton(
-                        onClick = {
-                            showSignInDialog = false
-                            signInDelaySeconds = 0
-                        }
-                    ) {
-                        Text(stringResource(android.R.string.cancel))
-                    }
-
-                    Button(
-                        onClick = {
-                            if (signInDelaySeconds == 0) {
-                                // Create anonymous account directly
-                                scope.launch {
-                                    val result = viewModel.webAuthManager.signInAnonymously()
-                                    if (result.isSuccess) {
-                                        // Sync preferences after successful authentication
-                                        viewModel.signInAnonymously()
-                                    }
-                                }
-                                showSignInDialog = false
-                                signInDelaySeconds = 0
-                            }
-                        },
-                        enabled = signInDelaySeconds == 0
-                    ) {
-                        Text(
-                            stringResource(
-                                if (signInDelaySeconds == 0) R.string.sync_account_create
-                                else R.string.sync_account_please_wait
-                            )
-                        )
-                    }
-                }
-            )
-        }
-    }
-}
-
-@Composable
-private fun FilterOptionCard(
-    title: String,
-    description: String,
-    isEnabled: Boolean,
-    onToggle: (Boolean) -> Unit,
-    icon: Int
-) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .border(
-                width = 1.5.dp,
-                color = MaterialTheme.colorScheme.primary.copy(alpha = 0.3f),
-                shape = RoundedCornerShape(12.dp)
-            ),
-        colors = CardDefaults.cardColors(
-            containerColor = if (isEnabled)
-                MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.4f)
-            else
-                MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.5f)
-        ),
-        shape = RoundedCornerShape(12.dp)
-    ) {
-        Column(modifier = Modifier.padding(14.dp)) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.SpaceBetween,
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    modifier = Modifier.weight(1f)
-                ) {
-                    Icon(
-                        painter = painterResource(icon),
-                        contentDescription = null,
-                        modifier = Modifier.size(24.dp),
-                        tint = MaterialTheme.colorScheme.primary
-                    )
-                    Spacer(modifier = Modifier.width(12.dp))
-                    Column {
-                        Text(
-                            text = title,
-                            style = MaterialTheme.typography.labelLarge.copy(fontWeight = FontWeight.SemiBold),
-                            color = MaterialTheme.colorScheme.onSurface
-                        )
-                        Text(
-                            text = description,
-                            style = MaterialTheme.typography.labelSmall,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                            modifier = Modifier.padding(top = 2.dp)
-                        )
-                    }
-                }
-
-                Switch(
-                    checked = isEnabled,
-                    onCheckedChange = onToggle,
-                    colors = SwitchDefaults.colors(
-                        checkedThumbColor = MaterialTheme.colorScheme.primary,
-                        checkedTrackColor = MaterialTheme.colorScheme.primary.copy(alpha = 0.3f),
-                        uncheckedThumbColor = MaterialTheme.colorScheme.outline,
-                        uncheckedTrackColor = MaterialTheme.colorScheme.outline.copy(alpha = 0.3f)
-                    )
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun PermissionsScreen(
-    onBack: () -> Unit,
-    onComplete: () -> Unit,
-) {
-    val context = LocalContext.current
-    val lifecycleOwner = LocalLifecycleOwner.current
-
-    var storageGranted by remember {
-        mutableStateOf(PermissionHelper.hasMediaStoreWritePermission(context))
-    }
-
-    var notificationsGranted by remember {
-        mutableStateOf(
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-                NotificationManagerCompat.from(context).areNotificationsEnabled()
-            } else {
-                true
-            }
-        )
-    }
-
-    var nearbyGranted by remember {
-        mutableStateOf(
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-                ContextCompat.checkSelfPermission(
-                    context,
-                    Manifest.permission.NEARBY_WIFI_DEVICES
-                ) == android.content.pm.PackageManager.PERMISSION_GRANTED
-            } else {
-                true
-            }
-        )
-    }
-
-    var backgroundGranted by remember { mutableStateOf(isIgnoringBatteryOptimizations(context)) }
-    var accessibilityGranted by remember { mutableStateOf(isAccessibilityEnabled(context)) }
-    var systemAlertGranted by remember {
-        mutableStateOf(
-            Settings.canDrawOverlays(context)
-        )
-    }
-    // PiP permission is declared in manifest but doesn't require runtime permission grant
-    // Mark as true by default since it's available on Android 8.0+
-    var pipGranted by remember { mutableStateOf(true) }
-
-    val storagePermissionLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.RequestMultiplePermissions()
-    ) { permissions ->
-        storageGranted = permissions.values.all { it }
-    }
-
-    val notificationLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.RequestPermission()
-    ) { granted ->
-        notificationsGranted = granted
-    }
-
-    val nearbyLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.RequestPermission()
-    ) { granted ->
-        nearbyGranted = granted
-    }
-
-    DisposableLifecycle(onEvent = {
-        backgroundGranted = isIgnoringBatteryOptimizations(context)
-        accessibilityGranted = isAccessibilityEnabled(context)
-        systemAlertGranted = Settings.canDrawOverlays(context)
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            notificationsGranted = NotificationManagerCompat.from(context).areNotificationsEnabled()
-            nearbyGranted = ContextCompat.checkSelfPermission(
-                context,
-                Manifest.permission.NEARBY_WIFI_DEVICES
-            ) == android.content.pm.PackageManager.PERMISSION_GRANTED
-        }
-        storageGranted = PermissionHelper.hasMediaStoreWritePermission(context)
-    }, lifecycleOwner = lifecycleOwner)
-
-    // Required permissions that must be granted to continue
-    val requiredGranted = storageGranted && notificationsGranted && backgroundGranted
-
-    val allGranted = listOf(
-        storageGranted,
-        notificationsGranted,
-        nearbyGranted,
-        backgroundGranted,
-        accessibilityGranted,
-        systemAlertGranted,
-        pipGranted
-    ).all { it }
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(MaterialTheme.colorScheme.surface)
-    ) {
-        Column(
-            verticalArrangement = Arrangement.spacedBy(18.dp),
-            modifier = Modifier
-                .align(Alignment.Center)
-                .fillMaxWidth(0.9f)
-        ) {
-                        Column(
-                verticalArrangement = Arrangement.spacedBy(4.dp),
-                modifier = Modifier.fillMaxWidth(),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Text(
-                    text = stringResource(R.string.onboarding_permissions_title),
-                    style = MaterialTheme.typography.headlineSmall.copy(fontWeight = FontWeight.Bold),
-                    textAlign = TextAlign.Center,
-                    color = MaterialTheme.colorScheme.primary
-                )
-                Text(
-                    text = if (allGranted) {
-                        stringResource(R.string.onboarding_all_set)
-                    } else {
-                        stringResource(R.string.onboarding_permissions_subtitle)
-                    },
-                    style = MaterialTheme.typography.bodySmall,
-                    color = if (allGranted) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.onSurfaceVariant,
-                    textAlign = TextAlign.Center
-                )
-            }
-
-            Spacer(modifier = Modifier.height(20.dp))
-
-            // Show only the first needed permission
-            if (!storageGranted) {
-                PermissionCard(
-                    title = stringResource(R.string.onboarding_perm_storage_title),
-                    description = stringResource(R.string.onboarding_perm_storage_desc),
-                    granted = storageGranted,
-                    actionLabel = stringResource(R.string.onboarding_grant),
-                ) {
-                    storagePermissionLauncher.launch(PermissionHelper.getRequiredWritePermissions())
-                }
-            } else if (!notificationsGranted) {
-                PermissionCard(
-                    title = stringResource(R.string.onboarding_perm_notifications_title),
-                    description = stringResource(R.string.onboarding_perm_notifications_desc),
-                    granted = notificationsGranted,
-                    actionLabel = stringResource(R.string.onboarding_grant),
-                ) {
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-                        notificationLauncher.launch(Manifest.permission.POST_NOTIFICATIONS)
-                    } else {
-                        openAppSettings(context)
-                    }
-                }
-            } else if (!backgroundGranted) {
-                PermissionCard(
-                    title = stringResource(R.string.onboarding_perm_background_title),
-                    description = stringResource(R.string.onboarding_perm_background_desc),
-                    granted = backgroundGranted,
-                    actionLabel = stringResource(R.string.onboarding_open_settings),
-                ) {
-                    openBatterySettings(context)
-                }
-            } else if (!accessibilityGranted) {
-                PermissionCard(
-                    title = stringResource(R.string.onboarding_perm_accessibility_title),
-                    description = stringResource(R.string.onboarding_perm_accessibility_desc),
-                    granted = accessibilityGranted,
-                    actionLabel = stringResource(R.string.onboarding_open_settings),
-                ) {
-                    openAccessibilitySettings(context)
-                }
-            } else if (!systemAlertGranted) {
-                PermissionCard(
-                    title = stringResource(R.string.onboarding_perm_system_alert_title),
-                    description = stringResource(R.string.onboarding_perm_system_alert_desc),
-                    granted = systemAlertGranted,
-                    actionLabel = stringResource(R.string.onboarding_open_settings),
-                ) {
-                    openSystemAlertSettings(context)
-                }
-            } else if (!nearbyGranted) {
-                PermissionCard(
-                    title = stringResource(R.string.onboarding_perm_nearby_title),
-                    description = stringResource(R.string.onboarding_perm_nearby_desc),
-                    granted = nearbyGranted,
-                    actionLabel = stringResource(R.string.onboarding_grant),
-                ) {
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-                        nearbyLauncher.launch(Manifest.permission.NEARBY_WIFI_DEVICES)
-                    } else {
-                        openAppSettings(context)
-                    }
-                }
-            } else if (!pipGranted) {
-                PermissionCard(
-                    title = stringResource(R.string.onboarding_perm_pip_title),
-                    description = stringResource(R.string.onboarding_perm_pip_desc),
-                    granted = pipGranted,
-                    actionLabel = stringResource(R.string.onboarding_grant),
-                ) {
-                    // PiP permission doesn't require explicit grant on most devices
-                    // Just mark as granted and continue
-                    pipGranted = true
-                }
-            }
-
-            Spacer(Modifier.height(16.dp))
-
-            Column(
-                modifier = Modifier.fillMaxWidth(),
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                if (!requiredGranted) {
-                    Text(
-                        text = stringResource(R.string.onboarding_permissions_required),
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.error,
-                        textAlign = TextAlign.Center,
-                        modifier = Modifier.fillMaxWidth()
-                    )
-                }
-
-                Button(
-                    onClick = onComplete,
-                    enabled = requiredGranted,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(44.dp),
-                    shape = RoundedCornerShape(10.dp),
-                    colors = ButtonDefaults.buttonColors(
-                        containerColor = MaterialTheme.colorScheme.primary
-                    )
-                ) {
-                    Text(
-                        text = stringResource(R.string.onboarding_continue),
-                        style = MaterialTheme.typography.labelMedium
-                    )
-                }
-
-                TextButton(
-                    onClick = onBack,
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    Text(
-                        text = stringResource(R.string.onboarding_back),
-                        style = MaterialTheme.typography.labelMedium,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun PermissionCard(
-    title: String,
-    description: String,
-    granted: Boolean,
-    actionLabel: String,
-    onAction: () -> Unit,
-) {
-    val indicatorColor by animateColorAsState(
-        targetValue = if (granted) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.outline,
-        label = "perm_indicator"
-    )
-
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .border(
-                width = 1.5.dp,
-                color = MaterialTheme.colorScheme.primary.copy(alpha = 0.3f),
-                shape = RoundedCornerShape(12.dp)
-            ),
-        colors = CardDefaults.cardColors(
-            containerColor = if (granted)
-                MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.4f)
-            else
-                MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.5f)
-        ),
-        shape = RoundedCornerShape(12.dp)
-    ) {
-        Column(modifier = Modifier.padding(14.dp)) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.SpaceBetween,
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Column(modifier = Modifier.weight(1f)) {
-                    Text(
-                        text = title,
-                        style = MaterialTheme.typography.labelLarge.copy(fontWeight = FontWeight.SemiBold),
-                        color = MaterialTheme.colorScheme.onSurface
-                    )
-                    Text(
-                        text = description,
-                        style = MaterialTheme.typography.labelSmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        modifier = Modifier.padding(top = 2.dp)
-                    )
-                }
-                val statusText = if (granted) {
-                    stringResource(R.string.onboarding_status_done)
-                } else {
-                    stringResource(R.string.onboarding_status_needed)
-                }
-                Box(
-                    modifier = Modifier
-                        .clip(CircleShape)
-                        .background(indicatorColor.copy(alpha = 0.12f))
-                        .padding(horizontal = 8.dp, vertical = 4.dp),
-                    contentAlignment = Alignment.Center
-                ) {
-                    Text(
-                        text = statusText,
-                        color = indicatorColor,
-                        style = MaterialTheme.typography.labelSmall.copy(fontWeight = FontWeight.Medium),
-                        textAlign = TextAlign.Center
-                    )
-                }
-            }
-
-            Spacer(Modifier.height(10.dp))
-
-            OutlinedButton(
-                onClick = onAction,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(38.dp),
-                shape = RoundedCornerShape(9.dp),
-                border = BorderStroke(1.5.dp, MaterialTheme.colorScheme.primary.copy(alpha = 0.7f)),
-                colors = ButtonDefaults.outlinedButtonColors(
-                    containerColor = MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.25f),
-                    contentColor = MaterialTheme.colorScheme.primary
-                )
-            ) {
-                Text(
-                    text = actionLabel,
-                    style = MaterialTheme.typography.labelSmall,
-                    color = MaterialTheme.colorScheme.primary
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun LegalOverlay(
-    title: String,
-    body: String,
-    onDismiss: () -> Unit,
-) {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.Black.copy(alpha = 0.4f))
-            .clickable { onDismiss() },
-        contentAlignment = Alignment.Center
-    ) {
-        Surface(
-            modifier = Modifier
-                .fillMaxWidth(0.88f)
-                .padding(20.dp),
-            shape = RoundedCornerShape(16.dp),
-            tonalElevation = 6.dp,
-            color = MaterialTheme.colorScheme.surface
-        ) {
-            Column(
-                modifier = Modifier
-                    .padding(20.dp)
-                    .verticalScroll(rememberScrollState()),
-                verticalArrangement = Arrangement.spacedBy(12.dp)
-            ) {
-                Text(
-                    text = title,
-                    style = MaterialTheme.typography.titleLarge.copy(fontWeight = FontWeight.Bold),
-                    color = MaterialTheme.colorScheme.primary
-                )
-                Text(
-                    text = body,
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-                Button(
-                    onClick = onDismiss,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(42.dp),
-                    shape = RoundedCornerShape(10.dp)
-                ) {
-                    Text(
-                        text = stringResource(R.string.ok),
-                        style = MaterialTheme.typography.labelMedium
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun LoadingScreen(
-    onFinished: () -> Unit,
-    shouldStartSync: Boolean = true,
-) {
-    val syncUtils = com.jtech.zemer.LocalSyncUtils.current
-    val progress by syncUtils.whitelistSyncProgress.collectAsState()
-    val composition by rememberLottieComposition(LottieCompositionSpec.RawRes(R.raw.loading_dots_blue))
-    val lottieColors = rememberLottieDynamicProperties(
-        rememberLottieDynamicProperty(
-            property = LottieProperty.COLOR_FILTER,
-            value = PorterDuffColorFilter(MaterialTheme.colorScheme.primary.toArgb(), PorterDuff.Mode.SRC_ATOP),
-            keyPath = arrayOf("**")
-        )
-    )
-
-    val loopingState = animateLottieCompositionAsState(
-        composition = composition,
-        iterations = LottieConstants.IterateForever,
-        restartOnPlay = true
-    )
-
-
-    LaunchedEffect(Unit) {
-        if (shouldStartSync) {
-            syncUtils.syncArtistWhitelist(forceSync = true)
-        }
-    }
-
-    LaunchedEffect(progress.isComplete) {
-        // Wait for sync to complete before finishing
-        if (progress.isComplete) {
-            onFinished()
-        }
-    }
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(MaterialTheme.colorScheme.surface),
-        contentAlignment = Alignment.Center
-    ) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center,
-            modifier = Modifier
-                .padding(horizontal = 24.dp, vertical = 24.dp)
-                .fillMaxWidth()
-        ) {
-            LottieAnimation(
-                composition = composition,
-                progress = { loopingState.progress },
-                dynamicProperties = lottieColors,
-                modifier = Modifier
-                    .size(320.dp)
-            )
-
-            Spacer(modifier = Modifier.height(48.dp))
-
-            Text(
-                text = stringResource(R.string.setting_up_library_title),
-                style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.Bold),
-                color = MaterialTheme.colorScheme.primary,
-                textAlign = TextAlign.Center,
-                modifier = Modifier.fillMaxWidth()
-            )
-        }
-    }
-}
-
-@Composable
-private fun DisposableLifecycle(
-    onEvent: () -> Unit,
-    lifecycleOwner: androidx.lifecycle.LifecycleOwner,
-) {
-    LaunchedEffect(lifecycleOwner) {
-        onEvent()
-    }
-
-    val observer = remember {
-        LifecycleEventObserver { _, event ->
-            if (event == Lifecycle.Event.ON_RESUME) {
-                onEvent()
-            }
-        }
-    }
-
-    DisposableEffectWithLifecycle(lifecycleOwner, observer)
-}
-
-@Composable
-private fun DisposableEffectWithLifecycle(
-    lifecycleOwner: androidx.lifecycle.LifecycleOwner,
-    observer: LifecycleEventObserver,
-) {
-    androidx.compose.runtime.DisposableEffect(lifecycleOwner) {
-        lifecycleOwner.lifecycle.addObserver(observer)
-        onDispose {
-            lifecycleOwner.lifecycle.removeObserver(observer)
-        }
-    }
-}
-
-@Composable
-private fun BottomNavSetupScreen(
-    onBack: () -> Unit,
-    onComplete: () -> Unit,
-) {
-    val context = LocalContext.current
-    var enableBottomNav by remember { mutableStateOf(true) }
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(MaterialTheme.colorScheme.surface)
-    ) {
-        Column(
-            verticalArrangement = Arrangement.spacedBy(14.dp),
-            modifier = Modifier
-                .align(Alignment.Center)
-                .fillMaxWidth(0.9f)
-        ) {
-                        Column(
-                verticalArrangement = Arrangement.spacedBy(4.dp),
-                modifier = Modifier.fillMaxWidth(),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Text(
-                    text = stringResource(R.string.onboarding_nav_setup),
-                    style = MaterialTheme.typography.headlineSmall.copy(fontWeight = FontWeight.Bold),
-                    textAlign = TextAlign.Center,
-                    color = MaterialTheme.colorScheme.primary
-                )
-                Text(
-                    text = stringResource(R.string.onboarding_nav_question),
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    textAlign = TextAlign.Center
-                )
-            }
-
-            Spacer(modifier = Modifier.height(20.dp))
-
-            // Choice cards — the shared onboarding radio card (D-pad focus treatment included)
-            Column(
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                OnboardingChoiceCard(
-                    selected = enableBottomNav,
-                    title = stringResource(R.string.onboarding_nav_enable),
-                    description = stringResource(R.string.onboarding_nav_enable_desc),
-                    onSelect = { enableBottomNav = true },
-                )
-                OnboardingChoiceCard(
-                    selected = !enableBottomNav,
-                    title = stringResource(R.string.onboarding_nav_no_thanks),
-                    description = stringResource(R.string.onboarding_nav_later),
-                    onSelect = { enableBottomNav = false },
-                )
-            }
-
-            Spacer(Modifier.height(16.dp))
-
-            Column(
-                modifier = Modifier.fillMaxWidth(),
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                Text(
-                    text = stringResource(R.string.onboarding_nav_customize_later),
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    textAlign = TextAlign.Center,
-                    modifier = Modifier.fillMaxWidth().padding(bottom = 4.dp)
-                )
-
-                Button(
-                    onClick = {
-                        // Save preference using SharedPreferences
-                        val prefs = context.getSharedPreferences("metrolist_settings", Context.MODE_PRIVATE)
-                        prefs.edit {
-                            putBoolean("bottomNavigationBarEnabled", enableBottomNav)
-                            // Set default items if enabling
-                            if (enableBottomNav) {
-                                putString("bottomNavigationItems", "home,search,library")
-                            }
-                        }
-                        onComplete()
-                    },
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(44.dp),
-                    shape = RoundedCornerShape(10.dp),
-                    colors = ButtonDefaults.buttonColors(
-                        containerColor = MaterialTheme.colorScheme.primary
-                    )
-                ) {
-                    Text(
-                        text = stringResource(R.string.onboarding_continue),
-                        style = MaterialTheme.typography.labelMedium
-                    )
-                }
-
-                TextButton(
-                    onClick = onBack,
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    Text(
-                        text = stringResource(R.string.onboarding_back),
-                        style = MaterialTheme.typography.labelMedium,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-            }
-        }
-    }
-}
-
-private fun isIgnoringBatteryOptimizations(context: Context): Boolean {
-    val pm = context.getSystemService(Context.POWER_SERVICE) as? PowerManager ?: return false
-    return pm.isIgnoringBatteryOptimizations(context.packageName)
-}
-
-@SuppressLint("BatteryLife", "UseKtx")
-private fun openBatterySettings(context: Context) {
-    runCatching {
-        val intent = Intent(Settings.ACTION_REQUEST_IGNORE_BATTERY_OPTIMIZATIONS).apply {
-            data = "package:${context.packageName}".toUri()
-            addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-        }
-        context.startActivity(intent)
-    }
-}
-
-private fun isAccessibilityEnabled(context: Context): Boolean {
-    val enabledServices = Settings.Secure.getString(
-        context.contentResolver,
-        Settings.Secure.ENABLED_ACCESSIBILITY_SERVICES
-    ) ?: return false
-    return enabledServices.contains("${context.packageName}/com.jtech.zemer.accessibility.ButtonMapperAccessibilityService")
-}
-
-private fun openAccessibilitySettings(context: Context) {
-    runCatching {
-        context.startActivity(
-            Intent(Settings.ACTION_ACCESSIBILITY_SETTINGS).addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-        )
-    }
-}
-
-@SuppressLint("UseKtx")
-private fun openSystemAlertSettings(context: Context) {
-    runCatching {
-        val intent = Intent(Settings.ACTION_MANAGE_OVERLAY_PERMISSION).apply {
-            data = "package:${context.packageName}".toUri()
-            addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-        }
-        context.startActivity(intent)
-    }
-}
-
-private fun openAppSettings(context: Context) {
-    val intent = Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS).apply {
-        data = Uri.fromParts("package", context.packageName, null)
-        addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-    }
-    context.startActivity(intent)
-}
