+++ app/src/main/kotlin/com/jtech/zemer/ui/screens/settings/AboutScreen.kt
@@ -1,6 +1,6 @@
 package com.jtech.zemer.ui.screens.settings
-import com.jtech.zemer.ui.component.AppNameTitle
 
+import com.jtech.zemer.ui.component.AppNameTitle
 import androidx.annotation.StringRes
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -21,7 +21,6 @@ import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.text.TextAutoSize
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.AssistChip
-import androidx.compose.material3.AssistChipDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
